Split node/builder
