+++ bee-node/src/logger.rs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::{node::BeeNodeBuilder, storage::Backend};
+
 use bee_common::logger::logger_init;
 
 impl<B: Backend> BeeNodeBuilder<B> {
